Remember password fo games
Fixes #1140
